Move services to generator module
