+++ src/app/modules/generator/components/fields/select-container/select-container.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { ContainerConfigField } from '../../../../../core/models/container';
+import { ContainerConfigField } from '../../../models/container';
 
 @Component({
     templateUrl: './select-container.component.html',
