+++ src/app/modules/generator/directives/form-field.directive.ts
@@ -1,11 +1,11 @@
 import { ComponentFactoryResolver, Directive, Input, OnInit, Type, ViewContainerRef } from '@angular/core';
 import { VoidComponent } from '../components/fields/void/void.component';
-import { ContainerConfigField } from '../../../core/models/container';
 import { InputComponent } from '../components/fields/input/input.component';
 import { CheckboxComponent } from '../components/fields/checkbox/checkbox.component';
 import { SelectSingleComponent } from '../components/fields/select-single/select-single.component';
 import { SelectMultipleComponent } from '../components/fields/select-multiple/select-multiple.component';
 import { SelectContainerComponent } from '../components/fields/select-container/select-container.component';
+import { ContainerConfigField } from '../models/container';
 
 @Directive({
     selector: '[appGeneratorFormField]'
