+++ src/app/modules/generator/interfaces/api-data.ts
