+++ src/app/modules/generator/interfaces/listable.ts
